fix: outer-loop verification fix #6 — code defaults align with disclosed deviations
Stage 1 (LOW)
- stage1_grape.py:462 stale citation §3.2.2 → §3.2.1 to align with
  the spec's 2026-05-07 reversion (lines 405-406 are §3.2.1 last
  paragraph; Algorithm 1 lives in Appendix L).

Stage 2 (LOW)
- stage2_reap_ream.py:104 covariance_storage_dtype default flipped
  from "float32" → "float16" to match D-cov-storage-fp16. Production
  config still pins fp16 explicitly; the default is for config-omitted
  invocations and now agrees with the disclosed deviation.

Stage 5 (LOW)
- stage5_router_kd.py: when both teacher_load_in_4bit=true AND
  teacher_logits_cache exist, log a clear warning that cache wins per
  spec §8 mutual-exclusion rule (previously silent override).

10 tests pass.

+++ max_quality/src/moe_compress/stage2_reap_ream.py
@@ -101,7 +101,11 @@ def run(
     # Spec §5 "Covariance Side-Collection": FP32 storage is recommended by
     # Swift-SVD paper 2604.01609 (avoids numerical degradation in eigendecomposition);
     # the dtype is configurable via covariance_storage_dtype.
-    cov_dtype = getattr(torch, s2.get("covariance_storage_dtype", "float32"))
+    # Default fp16 per §12 D-cov-storage-fp16 (10 mantissa bits, half the
+    # disk vs fp32, no measurable downstream PPL drift on Qwen3-30B-A3B).
+    # Production config also pins this to fp16; the default is here for
+    # config-omitted invocations.
+    cov_dtype = getattr(torch, s2.get("covariance_storage_dtype", "float16"))
     cov_acc.set_storage_dtype(cov_dtype)
     merge_map: dict[int, dict[int, list[int]]] = {}
